Drop Flink 1.17,1.18 support
