+++ gradle.properties
@@ -39,7 +39,7 @@ docker_image_default_repo_root=apache
 docker_image_default_repo_prefix=beam_
 
 # supported flink versions
-flink_versions=1.17,1.18,1.19,1.20,2.0,2.1
+flink_versions=1.19,1.20,2.0,2.1
 # supported spark versions
 spark_versions=3,4
 # supported python versions
