+++ sdks/go/examples/wasm/README.md
@@ -68,7 +68,7 @@ cd $BEAM_HOME
 Expected output should include the following, from which you acquire the latest flink runner version.
 
 ```shell
-'flink_versions: 1.17,1.18,1.19,1.20,2.0,2.1'
+'flink_versions: 1.19,1.20,2.0,2.1'
 ```
 
 #### 2. Set to the latest flink runner version i.e. 2.1
